Add files via upload
This program needs a database(.yml file) which contains the face data. This program won't run on itself. 
It needs a recognizer program which trains the data and creates a '.yml' file.
You will also need a lot of images of people you want to be detected by the program as the model needs a lot of data to improve the accuracy of the model.